pyproject.toml
[project]
name = "voice-french-agent"
version = "0.1.0"
description = "Real-time voice agent that listens to English speech and responds in French using AI translation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepgram-sdk>=4.1.0",
    "livekit-agents[deepgram,elevenlabs,mcp,openai,rime,silero]~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.0",
]
